project/backend/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-environ==0.11.1
django-grappelli==3.0.7
Pillow==10.0.0
sqlparse==0.4.4
tzdata==2023.3

psycopg2~=2.9.9
djangorestframework~=3.14.0
requests~=2.31.0
beautifulsoup4~=4.12.2
channels~=4.0.0